visual_entity_extraction/facenet_triton/mtcnn/requirements.txt
Flask==2.1.2
# Flask-Cors==3.0.10
numpy==1.22.3
Pillow==9.1.0
requests==2.27.1
